factor out path determination
